+++ CHANGELOG.md
@@ -6,7 +6,21 @@ The format is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.3.1 - 2024-10-03
+## 0.4.0 - 2025-02-05
+
+[View Changes on GitHub](https://github.com/jcayers20/mkdocs-autoapi/compare/0.3.2...0.4.0)
+
+### Bug Fixes
+
+- Fixed `mkdocs` crash if both `repo_url` and `edit_uri` were provided in
+`mkdocs.yml`
+- Fixed navigation when using the `mkdocs` theme
+
+### Features
+
+- Enabled VBA support
+
+## 0.3.2 - 2024-11-01
 
 [View Changes on GitHub](https://github.com/jcayers20/mkdocs-autoapi/compare/0.3.1...0.3.2)